common checkpath: use a 'double' temporary when swapping coordinates
It looks unintentional that this function was truncating to `int` when sorting
box points. This squashes 8 -Wfloat-conversion warnings.

+++ lib/common/routespl.c
@@ -690,7 +690,7 @@ static int checkpath(int boxn, boxf* boxes, path* thepath)
 	    printpath(thepath);
 	}
 	if (errs > 0) {
-	    int xy;
+	    double xy;
 
 	    if (l == 1)
 		xy = ba->UR.x, ba->UR.x = bb->LL.x, bb->LL.x = xy, l = 0;
